+++ packages/logging/service.go
@@ -10,6 +10,8 @@ import (
 	"github.com/cloudogu/cesapp-lib/registry"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/timestamppb"
+	v1 "k8s.io/api/apps/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"strings"
 	"time"
 
@@ -42,13 +44,18 @@ type doguDescriptionGetter interface {
 	GetCurrent(ctx context.Context, simpleDoguName string) (*core.Dogu, error)
 }
 
+type deploymentGetter interface {
+	Get(ctx context.Context, name string, opts metav1.GetOptions) (*v1.Deployment, error)
+}
+
 // NewLoggingService creates a new logging service.
-func NewLoggingService(provider logProvider, cp configProvider, restarter doguRestarter, descriptionGetter doguDescriptionGetter) *loggingService {
+func NewLoggingService(provider logProvider, cp configProvider, restarter doguRestarter, descriptionGetter doguDescriptionGetter, deploymentGetter deploymentGetter) *loggingService {
 	return &loggingService{
 		logProvider:          provider,
 		configProvider:       cp,
 		doguRestarter:        restarter,
 		doguDescriptorGetter: descriptionGetter,
+		deploymentGetter:     deploymentGetter,
 	}
 }
 
@@ -58,6 +65,7 @@ type loggingService struct {
 	configProvider       configProvider
 	doguRestarter        doguRestarter
 	doguDescriptorGetter doguDescriptionGetter
+	deploymentGetter     deploymentGetter
 }
 
 // QueryForDogu writes dogu log messages into the stream of the given server.
@@ -165,6 +173,16 @@ func (s *loggingService) setLogLevel(ctx context.Context, doguName string, l Log
 	}
 
 	logrus.Debugf("written new log level %s for dogu %s", l.String(), doguName)
+
+	deployment, err := s.deploymentGetter.Get(ctx, doguName, metav1.GetOptions{})
+	if err != nil {
+		return false, fmt.Errorf("could not get deployment to check status of dogu: %w", err)
+	}
+
+	if deployment.Spec.Replicas == nil || *deployment.Spec.Replicas == 0 {
+		logrus.Debugf("No replicas found for dogu %s, dogu is stopped", doguName)
+		return false, nil
+	}
 
 	return true, nil
 }