+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#31] After setting a new log level for a dogu, the dogu restarts even when it is stopped - dogu should stay stopped
 
 ## [v0.8.0] - 2024-05-23
 ### Changed